refactor(settings): move voice agent block under dictation agent tab
The 'Hey {{agentName}}' trigger is a dictation concern, not a chat
concern, so the agent-name UI now lives in DictationAgentSettings —
between the model picker and prompt customization when the dictation
agent is enabled, otherwise directly under the toggle. Storage key,
i18n keys, and useAgentName contract are unchanged.
