+++ src/components/SettingsPage.tsx
@@ -54,7 +54,6 @@ import {
 import { Alert, AlertTitle, AlertDescription } from "./ui/alert";
 import { useSettings } from "../hooks/useSettings";
 import { useDialogs } from "../hooks/useDialogs";
-import { useAgentName } from "../utils/agentName";
 import { useWhisper } from "../hooks/useWhisper";
 import { usePermissions } from "../hooks/usePermissions";
 import { useSystemAudioPermission } from "../hooks/useSystemAudioPermission";
@@ -726,8 +725,6 @@ export default function SettingsPage({
   const permissionsHook = usePermissions(showAlertDialog);
   const systemAudio = useSystemAudioPermission();
   useClipboard(showAlertDialog);
-  const { agentName, setAgentName } = useAgentName();
-  const [agentNameInput, setAgentNameInput] = useState(agentName);
   const [audioStorageUsage, setAudioStorageUsage] = useState<{
     fileCount: number;
     totalBytes: number;
@@ -782,38 +779,6 @@ export default function SettingsPage({
   useEffect(() => {
     refreshYdotoolStatus();
   }, [refreshYdotoolStatus]);
-
-  const handleSaveAgentName = useCallback(() => {
-    const trimmed = agentNameInput.trim();
-    const previousName = agentName;
-
-    setAgentName(trimmed);
-    setAgentNameInput(trimmed);
-
-    let nextDictionary = customDictionary.filter((w) => w !== previousName);
-    if (trimmed) {
-      const hasName = nextDictionary.some((w) => w.toLowerCase() === trimmed.toLowerCase());
-      if (!hasName) {
-        nextDictionary = [trimmed, ...nextDictionary];
-      }
-    }
-    setCustomDictionary(nextDictionary);
-
-    showAlertDialog({
-      title: t("settingsPage.agentConfig.dialogs.updatedTitle"),
-      description: t("settingsPage.agentConfig.dialogs.updatedDescription", {
-        name: trimmed,
-      }),
-    });
-  }, [
-    agentNameInput,
-    agentName,
-    customDictionary,
-    setAgentName,
-    setCustomDictionary,
-    showAlertDialog,
-    t,
-  ]);
 
   const { theme, setTheme } = useTheme();
   const usage = useUsage();
@@ -3069,108 +3034,7 @@ EOF`,
         return (
           <LlmsTabs
             initialTab={initialSubTab as LlmTab | undefined}
-            renderChatIntelligence={() => (
-              <div className="space-y-6">
-                <ChatAgentSettings />
-
-                <div className="border-t border-border/40 pt-6 space-y-5">
-                  <SectionHeader
-                    title={t("settingsPage.agentConfig.title")}
-                    description={t("settingsPage.agentConfig.description")}
-                  />
-
-                  <div>
-                    <p className="text-xs font-medium text-foreground mb-3">
-                      {t("settingsPage.agentConfig.agentName")}
-                    </p>
-                    <SettingsPanel>
-                      <SettingsPanelRow>
-                        <div className="space-y-3">
-                          <div className="flex gap-2">
-                            <Input
-                              placeholder={t("settingsPage.agentConfig.placeholder")}
-                              value={agentNameInput}
-                              onChange={(e) => setAgentNameInput(e.target.value)}
-                              className="flex-1 text-center text-base font-mono"
-                            />
-                            <Button
-                              onClick={handleSaveAgentName}
-                              disabled={!agentNameInput.trim()}
-                              size="sm"
-                            >
-                              {t("settingsPage.agentConfig.save")}
-                            </Button>
-                          </div>
-                          <p className="text-xs text-muted-foreground/60">
-                            {t("settingsPage.agentConfig.helper")}
-                          </p>
-                        </div>
-                      </SettingsPanelRow>
-                    </SettingsPanel>
-                  </div>
-
-                  <div>
-                    <SectionHeader title={t("settingsPage.agentConfig.howItWorksTitle")} />
-                    <SettingsPanel>
-                      <SettingsPanelRow>
-                        <p className="text-xs text-muted-foreground leading-relaxed">
-                          {t("settingsPage.agentConfig.howItWorksDescription", { agentName })}
-                        </p>
-                      </SettingsPanelRow>
-                    </SettingsPanel>
-                  </div>
-
-                  <div>
-                    <SectionHeader title={t("settingsPage.agentConfig.examplesTitle")} />
-                    <SettingsPanel>
-                      <SettingsPanelRow>
-                        <div className="space-y-2.5">
-                          {[
-                            {
-                              input: t("settingsPage.agentConfig.examples.formalEmail", {
-                                agentName,
-                              }),
-                              mode: t("settingsPage.agentConfig.instructionMode"),
-                            },
-                            {
-                              input: t("settingsPage.agentConfig.examples.professional", {
-                                agentName,
-                              }),
-                              mode: t("settingsPage.agentConfig.instructionMode"),
-                            },
-                            {
-                              input: t("settingsPage.agentConfig.examples.bulletPoints", {
-                                agentName,
-                              }),
-                              mode: t("settingsPage.agentConfig.instructionMode"),
-                            },
-                            {
-                              input: t("settingsPage.agentConfig.cleanupExample"),
-                              mode: t("settingsPage.agentConfig.cleanupMode"),
-                            },
-                          ].map((example, i) => (
-                            <div key={i} className="flex items-start gap-3">
-                              <span
-                                className={`shrink-0 mt-0.5 text-xs font-medium uppercase tracking-wider px-1.5 py-px rounded ${
-                                  example.mode === t("settingsPage.agentConfig.instructionMode")
-                                    ? "bg-primary/10 text-primary dark:bg-primary/15"
-                                    : "bg-muted text-muted-foreground"
-                                }`}
-                              >
-                                {example.mode}
-                              </span>
-                              <p className="text-xs text-muted-foreground leading-relaxed">
-                                "{example.input}"
-                              </p>
-                            </div>
-                          ))}
-                        </div>
-                      </SettingsPanelRow>
-                    </SettingsPanel>
-                  </div>
-                </div>
-              </div>
-            )}
+            renderChatIntelligence={() => <ChatAgentSettings />}
             renderDictationCleanup={() => (
               <div className="space-y-6">
                 <AiModelsSection
